# src/main/resources/openapi/api.yml
openapi: 3.1.0
info:
  title: "API Documentation"
  description: "API documentation for clientes-cuentas-microservicios"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
  /clientes:
    get:
      summary: "Obtener todos los clientes con sus cuentas bancarias"
      operationId: getClientes
      responses:
        '200':
          description: "Lista de todos los clientes"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteResponse'

  /clientes/{dni}:
    get:
      summary: "Obtener cliente por DNI"
      operationId: getClientesPorDni
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Cliente encontrado"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'

  /clientes/mayores-de-edad:
    get:
      summary: "Obtener todos los clientes mayores de edad"
      operationId: getClientesMayoresDeEdad
      responses:
        '200':
          description: "Lista de clientes mayores de edad"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteResponse'

  /clientes/con-cuenta-superior-a/{cantidad}:
    get:
      summary: "Obtener clientes con cuentas que superan cierta cantidad"
      operationId: getClientesConCuentaSuperiorA
      parameters:
        - name: cantidad
          in: path
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: "Lista de clientes que cumplen el criterio"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteResponse'

  /cuentas:
    post:
      summary: "Crear una nueva cuenta bancaria"
      operationId: createCuentaBancaria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCuentaBancariaRequest'
      responses:
        '201':
          description: "Cuenta creada exitosamente"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaBancariaResponse'

  /cuentas/{id}:
    put:
      summary: "Actualizar el total de una cuenta bancaria"
      operationId: updateCuentaBancaria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCuentaBancariaRequest'
      responses:
        '200':
          description: "Cuenta actualizada correctamente"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaBancariaResponse'

components:
  schemas:
    ClienteResponse:
      type: object
      properties:
        dni:
          type: string
        nombre:
          type: string
        apellido1:
          type: string
        apellido2:
          type: string
        fechaNacimiento:
          type: string
          format: date
        cuentas:
          type: array
          items:
            $ref: '#/components/schemas/CuentaBancariaResponse'

    CuentaBancariaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dniCliente:
          type: string
        tipoCuenta:
          type: string
          enum:
            - JUNIOR
            - NORMAL
            - PREMIUM
        total:
          type: number
          format: double

    CreateCuentaBancariaRequest:
      type: object
      properties:
        dniCliente:
          type: string
        tipoCuenta:
          type: string
          enum:
            - JUNIOR
            - NORMAL
            - PREMIUM
        total:
          type: number
          format: double
      required:
        - dniCliente
        - tipoCuenta
        - total

    UpdateCuentaBancariaRequest:
      type: object
      properties:
        total:
          type: number
          format: double
      required:
        - total